refactor: create use fetch gists hook
