+++ src/containers/gists/Gists.tsx
@@ -1,6 +1,6 @@
 import React, {useEffect} from 'react'
 import {withNamespaces} from 'react-i18next'
-import {useDispatch, useSelector} from 'react-redux'
+import {useSelector} from 'react-redux'
 
 import Headings from '../../shared/components/headings/Headings'
 import SeeMore from '../../shared/components/see-more/SeeMore'
@@ -8,49 +8,35 @@ import Loading from '../../shared/components/loading/Loading'
 import {ReactComponent as GitHubLogo} from '../../assets/images/social/github.svg'
 import {settings} from '../../settings'
 
-import * as actions from './redux/GistsActions'
 import {GistsState} from './redux/GistsReducer'
 
 import styles from './Gists.module.css'
-import {gistsService} from './services/GistsService'
 import Message from './components/Message'
 import GistsList from './components/GistsList'
 import {domService} from '../../shared/services/DOMService'
 
+import {useFetchGists} from './hooks/UseGists.hook'
+
 const Gists = ({t}: any) => {
-  const dispatch = useDispatch()
+  const fetchGists = useFetchGists()
   const gistsState: GistsState = useSelector((state: any) => state.gists)
 
   useEffect(() => {
     let ignore = false
 
-    const fetchGists = async () => {
-      dispatch(actions.fetchGistsInit())
-
-      try {
-        if (!ignore) {
-          const response = await gistsService.fetchGists()
-          const gistData = await gistsService.save(response)
-
-          setTimeout(() => {
-            dispatch(actions.fetchGistsSuccess(gistData))
-          }, settings.loading.delay)
-        }
-      } catch (error) {
-        dispatch(actions.fetchGistsFailure(error))
-      }
+    if (!ignore) {
+      domService.observeSection({
+        element: document.getElementById('section-gists'),
+        threshold: settings.loading.threshold,
+        callback: fetchGists
+      })
     }
-
-    domService.observeSection({
-      element: document.getElementById('section-gists'),
-      threshold: settings.loading.threshold,
-      callback: fetchGists
-    })
 
     return () => {
       ignore = true
     }
-  }, [dispatch])
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [])
 
   return (
     <section id="section-gists" className={`section ${styles['section-gists']}`}>
